Moved temp type file back to profile util folder
