+++ src/app/profile/components/base-form.tsx
@@ -15,7 +15,7 @@ import {
 import { Button } from '~/components/ui/button';
 import { Input } from '~/components/ui/input';
 import type { AllFormValues } from '../utils/form-schemas';
-import type { FormControlObj } from '../../../lib/types';
+import type { FormControlObj } from '../utils/types';
 
 interface BaseFormProps {
   submitFunction: (values: AllFormValues) => void;
